+++ helpers/copyFile.js
