+++ main.js
@@ -5,8 +5,9 @@ import path from 'node:path';
 import { readFilebyStream } from './helpers/readFileByStream.js';
 import { createFile } from './helpers/createFile.js';
 import { renameFile } from './helpers/renameFile.js';
-import { copyFile } from './copyFile.js';
+import { copyFile } from './helpers/copyFile.js';
 import { deleteFile } from './helpers/deleteFile.js';
+import {moveFile} from './helpers/moveFile.js';
 let currentPath = os.userInfo().homedir
 const userNameArg = process.argv.find((arg) => arg.startsWith('--username='))
 let username = null
@@ -78,21 +79,25 @@ rl.on('line', async (line) => {
         } catch(err){
             console.log(`Error: ${err.message}`)
         }
-    } else if(trimmedLine.startsWith('cat ')){
+    } else if(trimmedLine.startsWith('cat')){
         const filePath = trimmedLine.split(' ')[1]
         readFilebyStream(filePath)
 
-    } else if(trimmedLine.startsWith('add ')){
+    } else if(trimmedLine.startsWith('add')){
         const filePath = trimmedLine.split(' ')[1]
         createFile(filePath)
 
-    } else if (trimmedLine.startsWith('rn ')){
+    } else if (trimmedLine.startsWith('rn')){
        const [_,oldName,newName] = trimmedLine.split(' ')
        await renameFile(oldName,newName)
 
-    } else if (trimmedLine.startsWith('cp ')){
+    } else if (trimmedLine.startsWith('cp')){
         const [_,filePath, newFilePath] = trimmedLine.split(' ')
         await copyFile(filePath,newFilePath)
+
+    } else if (trimmedLine.startsWith('mv')){
+        const [_,filePath, newFilePath] = trimmedLine.split(' ')
+        await moveFile(filePath,newFilePath)
 
     } else if (trimmedLine.startsWith('rm')){
         const filePath = trimmedLine.split(' ')[1]
